Bug Fixes.
In omega/ImageUtils.cpp, ImageUtils::getPreallocatedBlock returned a false to a void*, which was throwing errors. It now returns a NULL as 'nullptr' is not avaliable.

In several other areas, an enum in EventBase to a returned enum of type Service::ServiceType were being compared. This originates from the two enums using equivalent values but being of different types in omicron's source. This was fixed with a static_cast to the EventBase side.